Todos los id pasados de Int a Long, y en teoría se insertan las especialidades del usuario

+++ LaboralEX/app/src/main/java/com/example/laboralex/database/dao/CompanyDao.kt
@@ -18,5 +18,5 @@ interface CompanyDao {
     suspend fun getAll(): List<Company>
 
     @Query("SELECT * FROM Company WHERE id = :id")
-    suspend fun getById(id: Int): Company
+    suspend fun getById(id: Long): Company
 }
