+++ LaboralEX/app/src/main/java/com/example/laboralex/database/entity/User.kt
@@ -6,8 +6,8 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class User(
-    @PrimaryKey(autoGenerate = true) val uid: Int = 0,
+    @PrimaryKey(autoGenerate = true) val uid: Long = 0,
     @ColumnInfo(name = "first_name") val firstName: String?,
     @ColumnInfo(name = "surnames") val surnames: String?,
-    @ColumnInfo(name = "profile_picture") val profilePictureId: Int?
+    @ColumnInfo(name = "profile_picture") val profilePictureId: Long?
 )
